OMS-19 update and delete activity service methods added

+++ src/services/activityService.ts
@@ -48,16 +48,42 @@ export const createActivity = async (activity: Partial<Activity>) => {
     }
 }
 
-export const updateActivity = async (id: string, activity: Activity) => {
-    const response = await axios.put(`${API_URL}/${id}`, activity, {
-        headers: { Authorization: `Bearer ${getToken()}` }
-    });
-    return response.data;
+export const updateActivity = async (activity: Partial<Activity>) => {
+    const token = getToken();
+    if (token) {
+        try {
+            const response = await axios.patch(`${API_URL}/${activity.id}`, activity, {
+                headers: {
+                    "Content-Type": 'application/json',
+                    Authorization: `Bearer ${token}`
+                }
+            });
+            return response.data;
+        } catch (error) {
+            console.error('An error occurred:', error);
+            if (axios.isAxiosError(error)) {
+                console.error('Error response:', error.response?.data);
+            }
+        }
+    }
 }
 
 export const deleteActivity = async (id: string) => {
-    const response = await axios.delete(`${API_URL}/${id}`, {
-        headers: { Authorization: `Bearer ${getToken()}` }
-    });
-    return response.data;
+    const token = getToken();
+    if (token) {
+        try {
+            const response = await axios.delete(`${API_URL}/${id}`, {
+                headers: {
+                    "Content-Type": 'application/json',
+                    Authorization: `Bearer ${token}`
+                }
+            });
+            return response.data;
+        } catch (error) {
+            console.error('An error occurred:', error);
+            if (axios.isAxiosError(error)) {
+                console.error('Error response:', error.response?.data);
+            }
+        }
+    }
 }
